trying to acheive modularity and implement tostring
